requirements.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.8.1
flake8==3.9.2
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.0
mccabe==0.6.1
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.9
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytils==0.3
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
